:bug: Fixed tests
