+++ api/test/tests_routes/Metrics.ts
@@ -1,0 +1,20 @@
+import type { InjectOptions } from 'fastify';
+import type { server } from '../utils/testsType';
+
+export async function METRICS_ENDPOINTS(server: server) {
+  describe('METRICS ENDPOINTS', () => {
+    const { app, expectedObjects } = server;
+    const inject: Record<string, InjectOptions> = {
+      getGlobalMetrics: { 
+        method: 'GET',
+        url: '/metrics',
+      },
+    };
+
+    test('GET GLOBAL METRICS', async () => {
+      const res = await app.inject(inject.getGlobalMetrics);
+      expect(res.statusCode).toEqual(200);
+      expect(await res.json()).toEqual(expectedObjects.metrics);
+    });
+  });
+}
